Adding query for active monitor's desktops
Provides a way to query `chunkc` for the collection of destops for the currently active monitor.

Response is in the format:

    chunkc tiling::query --monitor desktops
    2:13 2:8 2:11 2:10 2:12 2:9

Where the list of desktops is `monitorID:desktopID` space deliniated.

The addition of the monitor id to the output saves an additional request to `chunkc` for this information. This format fit my use case, but I would not object to the return from this query being changed to omit the monitor ID.